requirements.txt
pygame==2.1.3
requests==2.28.2
pycryptodome==3.17
cryptography==3.4.8